debugging

+++ windows/install.ps1
@@ -1,30 +1,51 @@
 param(
-    [Parameter(Mandatory = $false)][switch]$WhatIf
+    [Parameter(Mandatory = $false)][switch]$WhatIf,
+    [Parameter(Mandatory = $false)][switch]$Debug
 )
 
-$Definitions = & "$PSScriptRoot/Get-Scoopfile.ps1"
+if(!$IsWindows -and !$WhatIf) {
+    throw "Cannot actually install components on non-Windows platforms."
+}
 
-$Definitions | ForEach-Object {
-    $name = $_.Name
-    $isInstalled = $false
-    if ($_.Check -ne $null) {
-        $isInstalled = & $_.Check $_
+$oldDebugPref = $DebugPreference
+try {
+    if($Debug) {
+        $DebugPreference = "Continue"
     }
 
-    if ($isInstalled) {
-        Write-Host "Using $name"
-    }
-    else {
-        Write-Host -ForegroundColor Green "Installing $name ..."
-        if ($_.Install -eq $null) {
-            Write-Error "No install command defined for $name!"
+    Write-Debug "Loading scoopfile..."
+    $Definitions = & "$PSScriptRoot/Get-Scoopfile.ps1"
+
+    Write-Debug "Running items..."
+    $Definitions | ForEach-Object {
+        $name = $_.Name
+        Write-Debug "Processing $name"
+        $isInstalled = $false
+        if ($_.Check -ne $null) {
+            Write-Debug "Running checker for $name ..."
+            $isInstalled = & $_.Check $_
+        } else {
+            Write-Debug "No checker for $name!"
         }
-        elseif (!$WhatIf) {
-            try {
-                & $_.Install $_
-            } catch {
-                Write-Error "Installing $name failed with $($error[0])"
+
+        if ($isInstalled) {
+            Write-Host "Using $name"
+        }
+        else {
+            Write-Host -ForegroundColor Green "Installing $name ..."
+            if ($_.Install -eq $null) {
+                Write-Error "No install command defined for $name!"
+            }
+            elseif (!$WhatIf) {
+                try {
+                    Write-Debug "Running installer for $name ..."
+                    & $_.Install $_
+                } catch {
+                    Write-Error "Installing $name failed with $($error[0])"
+                }
             }
         }
     }
+} finally {
+    $DebugPreference = $oldDebugPref
 }
